issue #158: dedicated models module
